Atomic structure is implemented and Components folder is restructured. CSS changed in order to make app look 100% similar
