--- components/footer.js
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-
-export const FooterGrayStripe = styled.div`
-    width: 100%;
-    height: 35px;
-    background-color: #e8e6e6;
-    margin: 40px auto;
-    padding-left : 20%; 
-    
-`
