+++ components/organisms/footer/Footer.jsx
@@ -1,0 +1,7 @@
+import styles from './footer.module.scss';
+
+export default function Footer() {
+    return (
+        <p className = {styles.footer}> Copyright <span>&copy;</span> 2011-2018 Sabka Bazaar Grocery Supplies Pvt Ltd </p>
+    )
+}
