+++ components/organisms/footer/footer.module.scss
@@ -1,0 +1,7 @@
+.footer {
+    width: 100%;
+    height: 35px;
+    background-color: #e8e6e66b;
+    margin: 40px auto;
+    padding-left : 20%; 
+}
