--- components/products/productsList.module.css
@@ -1,4 +1,0 @@
-.productsList {
-       height: 50px;
-}
-
